Refactoring: Split fragment object retrieval and converstion
Fragment object retrieval is now moved to 'AasxServerStandardBib' as it can be implemented as a generic service.

Fragment object conversion resides in 'Swagger.Lib.V3' as it relies on the model definition for Content, Level, and Extend to convert a fragment object to a suitable representation for the HTTP response.